refactor: modularize codebase into sub-packages
- Split monolithic azure_api.py (1189 lines) into azure_api/ package:
  _auth.py, _cache.py, _pagination.py, discovery.py, skus.py,
  quotas.py, spot.py, pricing.py
- Extract shared evaluation helpers into services/_evaluation_helpers.py
  (GPU_FAMILY_MARKERS, SPOT_RANK, DATA_RESIDENCY_REGIONS, etc.)
- Move scoring functions to scoring/ package
  (admission_confidence.py, fragmentation.py)
- Group signal pipeline into services/signals/ subpackage
  (signal_store.py, signal_collector.py, volatility.py)
- Remove dead code (capacity_confidence.py, test_scoring_iso.py)
- Update all mock targets in tests for new module paths
- Backward-compatible re-exports in __init__.py and wrapper modules
- All 303 tests pass, mypy strict clean, ruff clean
