+++ src/az_scout/services/signals/__init__.py
@@ -1,0 +1,1 @@
+"""Signal pipeline – time-series collection, storage, and volatility analysis."""
